tenant-resolver-service/database_service.py
"""
Service for managing tenant databases.
"""

import os
import psycopg2
from models import TenantModel

def provision_database(tenant: TenantModel):
    """
    Provision a new database for a tenant.
    """
    # TODO: Implement database provisioning logic
    # 1. Connect to PostgreSQL server
    # 2. Create a new database for the tenant
    # 3. Create a database user with limited permissions
    # 4. Grant permissions to the user
    pass

def run_migrations(tenant: TenantModel):
    """
    Run database migrations for a tenant.
    """
    # TODO: Implement migration runner for tenant database
    # 1. Connect to tenant database
    # 2. Apply migrations using Alembic
    pass